Replace Legacy tab peek in new code
